requirements.txt
tesseract
PyPDF2
pytesseract
pillow
PyMuPDF
transformers
torch
streamlit
requests
sacremoses
